# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/owners": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all owners",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "owners"
                ],
                "summary": "Get all owners",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 15,
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.MultipleDataResp-owner_Owner"
                        }
                    }
                }
            },
            "post": {
                "description": "Add new owner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "owners"
                ],
                "summary": "Add new owner",
                "parameters": [
                    {
                        "description": "Add request",
                        "name": "owner",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/owner.AddReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SingleDataResp-owner_Owner"
                        }
                    }
                }
            }
        },
        "/owners/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get owner by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "owners"
                ],
                "summary": "Get owner by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "owner id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SingleDataResp-owner_Owner"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete owner by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "owners"
                ],
                "summary": "Delete owner by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "owner id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SingleDataResp-owner_Owner"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update owner by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "owners"
                ],
                "summary": "Update owner by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "owner id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update request",
                        "name": "owner",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/owner.UpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SingleDataResp-owner_Owner"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.MultipleDataResp-owner_Owner": {
            "type": "object",
            "required": [
                "data",
                "message",
                "pagination"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "x-order": "1"
                },
                "pagination": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.Pagination"
                        }
                    ],
                    "x-order": "2"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/owner.Owner"
                    },
                    "x-order": "2"
                }
            }
        },
        "api.Pagination": {
            "type": "object",
            "required": [
                "page",
                "pageSize",
                "totalItem",
                "totalPage"
            ],
            "properties": {
                "page": {
                    "type": "integer",
                    "x-order": "1"
                },
                "pageSize": {
                    "type": "integer",
                    "x-order": "2"
                },
                "totalPage": {
                    "type": "integer",
                    "x-order": "3"
                },
                "totalItem": {
                    "type": "integer",
                    "x-order": "4"
                }
            }
        },
...
        "api.SingleDataResp-owner_Owner": {
            "type": "object",
            "required": [
                "data",
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "x-order": "1"
                },
                "data": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/owner.Owner"
                        }
                    ],
                    "x-order": "2"
                }
            }
        },
...
        "owner.AddReq": {
            "type": "object",
            "required": [
                "images",
                "name"
            ],
            "properties": {
                "images": {
                    "type": "string",
                    "x-order": "1"
                },
                "name": {
                    "type": "string",
                    "x-order": "2"
                }
            }
        },
        "owner.Owner": {
            "type": "object",
            "required": [
                "createdAt",
                "id",
                "images",
                "name",
                "updatedAt"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "x-order": "1"
                },
                "images": {
                    "type": "string",
                    "x-order": "2"
                },
                "name": {
                    "type": "string",
                    "x-order": "3"
                },
                "createdAt": {
                    "type": "string",
                    "x-order": "4"
                },
                "updatedAt": {
                    "type": "string",
                    "x-order": "5"
                }
            }
        },
        "owner.UpdateReq": {
            "type": "object",
            "properties": {
                "images": {
                    "type": "string",
                    "x-order": "1"
                },
                "name": {
                    "type": "string",
                    "x-order": "2"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}